Refactor telemetry types into cohesive modules
Extracted `telemetry/src/types.rs` into `trace.rs`, `log.rs`, `meta.rs`, `wire.rs`, and `metrics_types.rs` to improve cohesion.
`types.rs` is now a facade re-exporting these modules to maintain backward compatibility.
Updated internal imports in `userspace/layer.rs`, `userspace/metrics.rs`, and `kernel/ring_buffer.rs`.
Verified `no_std` compatibility.
